backend/requirements.txt
# Django core
Django>=5.1,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.0
djangorestframework-simplejwt>=5.3

# ASGI & WebSocket
daphne>=4.1
channels>=4.0
channels-redis>=4.2

# Database
psycopg2-binary>=2.9
dj-database-url>=2.1
pgvector>=0.3

# Background tasks
celery>=5.4
redis>=5.0

# Scientific / Signal Processing
numpy>=1.26
scipy>=1.12
tables>=3.9
scikit-learn>=1.4
h5py>=3.10

# Agent framework
fastapi>=0.115
uvicorn>=0.30
httpx>=0.27

# LLM / Agentic
langgraph>=0.2
langchain-core>=0.3
langchain-ollama>=0.2
langfuse>=2.0

# MCP & A2A
mcp>=1.0

# Utilities
python-dotenv>=1.0
pydantic>=2.5
pydantic-settings>=2.1
python-dateutil>=2.9
orjson>=3.9
gunicorn>=22.0
whitenoise>=6.6
